Moving regex decls to a separate header
Moving the class declarations for util classes to parse and
evaluate user-specified regular expressions to save I/O
decompositions to a separate header file.

Code refactoring, no new code added.
