+++ src/clib/pio_sdecomps_regex.h
@@ -1,6 +1,6 @@
-#ifndef __PIO_SDECOMPS_REGEX__
-#define __PIO_SDECOMPS_REGEX__
+#ifndef __PIO_SDECOMPS_REGEX_H___
+#define __PIO_SDECOMPS_REGEX_H___
 
 bool pio_save_decomps_regex_match(int ioid, const char *fname, const char *vname);
 
-#endif /* __PIO_SDECOMPS_REGEX__ */
+#endif /* __PIO_SDECOMPS_REGEX_H__ */
